Added interface Add, clone and delete functionalities. Fixed some bugs with adding and refreshing. General or interface config missing.
Changed connection of comboboxes and lineedits to respond to user and not to the program for it to not generate errors on refreshing.
Changed init in CH to respond to the 3 databases, added some functions.
Added required data init function.

To do:
Add infos, types restrictions to Interface config.
Add required data from main
Prio 1: fix streaming add combo for type of streaming

Remove unused fuctions from multi
Add configs to files and test Marlin config