Added General configuration Testing and type checking
Removed interfaceId_info and interfaceId_type, were causing problems having them in 3 files. When present the interface add and delete does not add this two and causes errors.
user cannot change action with interfaceId in the text

CH changed the way it loads, for it to use different sets of data.

To do:
-set required data for the main interface correctly (reflects on how much you can change in the configuration through the program)
-Do a refresh after delete interface
-Fix streaming (add combo)
-xyz_update with multiinterface
-clean xyz multiinterface
-machine configuration read with multiinterface

-add external programs -gcode ripper and others->gcode visualization

-add loops and scripting (json maybe?)