VBD added batton cloning Added window settings to config file ...no idea how to install/reinstall window settings still... but is stored in btncfg files does not work as normal since btncfg files are json and object do not translate.
Added QuitreadOnMatch in configuration, to stop reading when read match is found. In case of tiny g was matching 1 word and missing all other parts in message.
Messages come separated or together as posx:0.000,Posy:10.0 on diagonal or posx:0.00,vel:600.0,stat:5.
was only reading the stat because it came first in the list.